changed random integers to floats

+++ app.rb
@@ -31,8 +31,8 @@ get("/random/new") do
 end
 
 get("/random/results") do
-  @min_number = params.fetch("user_min").to_i
-  @max_number = params.fetch("user_max").to_i
+  @min_number = params.fetch("user_min").to_f
+  @max_number = params.fetch("user_max").to_f
   @rand_number = rand(@min_number..@max_number)
   
   erb(:random_number_results)
